fix: 피드백 반영 - 도메인의 View 역할을 별도의 담당 클래스로 분리
