+++ src/main/kotlin/racingcar/domain/Car.kt
@@ -15,10 +15,6 @@ data class Car(
         }
     }
 
-    override fun toString(): String {
-        return "$name : " + CAR_LOCATION_PATH_SYMBOL.repeat(location)
-    }
-
     private fun isMovablePower(power: Int): Boolean =
         power in MOVABLE_MIN_POWER..MOVABLE_MAX_POWER
 
@@ -26,7 +22,6 @@ data class Car(
         const val CAR_DEFAULT_LOCATION = 0
         private const val MOVABLE_MAX_POWER = 9
         private const val MOVABLE_MIN_POWER = 4
-        private const val CAR_LOCATION_PATH_SYMBOL = "-"
         private const val CAR_MAX_NAME_LENGTH = 5
     }
 }
