+++ src/main/kotlin/racingcar/domain/RacingGameResult.kt
@@ -6,15 +6,4 @@ class RacingGameResult(val cars: List<Car>) {
         val maxLocation = cars.maxBy { it.location }.location
         return cars.filter { it.location == maxLocation }
     }
-
-    override fun toString(): String {
-        val stringBuilder: StringBuilder = StringBuilder()
-
-        for (car in cars) {
-            stringBuilder.append(car)
-            stringBuilder.append("\n")
-        }
-
-        return stringBuilder.toString()
-    }
 }
