+++ src/main/kotlin/racingcar/view/ResultView.kt
@@ -7,7 +7,7 @@ private const val RESULT_MESSAGE = "실행 결과"
 fun printRacingGameResult(racingGameResults: List<RacingGameResult>) {
     println(RESULT_MESSAGE)
     for (racingGameResult in racingGameResults) {
-        println(racingGameResult)
+        println(racingGameResult.state())
     }
 
     val winnersName = racingGameResults.last().getWinners().joinToString(separator = ", ") { it.name }
